+++ plugins/mesh/haze/object/haze.h
@@ -357,7 +357,7 @@ public:
   { this->directional=pos;}
   virtual const csVector3& GetDirectional() const
   {return this->directional;}
-  virtual int GetLayerCount() const {return this->layers.Length();}
+  virtual size_t GetLayerCount() const {return this->layers.Length();}
   virtual void AddLayer(iHazeHull *hull, float scale)
   {
     csHazeLayer *lay = new csHazeLayer(hull, scale);
@@ -435,7 +435,7 @@ public:
   { this->directional=pos;}
   virtual const csVector3& GetDirectional() const
   {return this->directional;}
-  virtual int GetLayerCount() const {return this->layers.Length();}
+  virtual size_t GetLayerCount() const {return this->layers.Length();}
   virtual void AddLayer(iHazeHull *hull, float scale)
   {
     csHazeLayer *lay = new csHazeLayer(hull, scale);
